changed datatype

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -8,7 +8,7 @@
  */
 size_t list_len(const list_t *h)
 {
-	int elements;
+	size_t elements;
 
 	while (h != NULL)
 	{
